Added more higher-level functions, minor fixes/improvements
Added a bunch more higher-level get/set functions. These need to be tested. Also re-ordered these functions in the headers to better match their order in terms of the address layout.

Address #defines are now uint8 type. This was required for the 0 address not to be ambiguous, so the rest were converted too. Might save a small amount of memory too depending on what the compiler does.

+++ DynamixelQ/DXL_AX.h
@@ -2,7 +2,7 @@
  *	DXL_AX.h
  *
  *	Author: Andrew D. Horchler, adh9 @ case.edu
- *	Created: 8-13-14, modified: 10-3-14
+ *	Created: 8-13-14, modified: 10-22-15
  */
 
 #ifndef DXL_AX_H_
@@ -22,9 +22,9 @@
 #define DXL_AX_DEGREES_PER_TICK (360.0/DXL_AX_POSITION_RESOLUTION)
 
 // Dynamixel register addresses for AX series
-#define DXL_CW_COMPLIANCE_MARGIN 26
-#define DXL_CCW_COMPLIANCE_MARGIN 27
-#define DXL_CW_COMPLIANCE_SLOPE 28
-#define DXL_CCW_COMPLIANCE_SLOPE 29
+#define DXL_CW_COMPLIANCE_MARGIN uint8(26)
+#define DXL_CCW_COMPLIANCE_MARGIN uint8(27)
+#define DXL_CW_COMPLIANCE_SLOPE uint8(28)
+#define DXL_CCW_COMPLIANCE_SLOPE uint8(29)
 
 #endif /* DXL_AX_H_ */
