+++ DynamixelQ/DXL_MX.h
@@ -2,7 +2,7 @@
  *	DXL_MX.h
  *
  *	Author: Andrew D. Horchler, adh9 @ case.edu
- *	Created: 8-13-14, modified: 10-3-14
+ *	Created: 8-13-14, modified: 10-22-15
  */
 
 #ifndef DXL_MX_H_
@@ -25,21 +25,21 @@
 #define DXL_MX_DEGREES_PER_TICK (360.0/DXL_MX_POSITION_RESOLUTION)
 
 // Dynamixel register addresses for MX series
-#define DXL_MULTI_TURN_OFFSET 20
-#define DXL_RESOLUTION_DIVIDER 22
+#define DXL_MULTI_TURN_OFFSET uint8(20)
+#define DXL_RESOLUTION_DIVIDER uint8(22)
 
 #define DXL_MULTI_TURN_MAX_OFFSET_MAGNITUDE 0x6000
 #define DXL_MULTI_TURN_MAX_RESOLUTION_DIVIDER 0x04
 
-#define DXL_D_GAIN 26
-#define DXL_I_GAIN 27
-#define DXL_P_GAIN 28
+#define DXL_D_GAIN uint8(26)
+#define DXL_I_GAIN uint8(27)
+#define DXL_P_GAIN uint8(28)
 
 // DXL_MX64 and DXL_MX106 only
-#define DXL_CURRENT 68
-#define DXL_TORQUE_CONTROL_MODE_ENABLE 69
-#define DXL_GOAL_TORQUE 71
+#define DXL_CURRENT uint8(68)
+#define DXL_TORQUE_CONTROL_MODE_ENABLE uint8(69)
+#define DXL_GOAL_TORQUE uint8(71)
 
-#define DXL_GOAL_ACCELERATION 73
+#define DXL_GOAL_ACCELERATION uint8(73)
 
 #endif /* DXL_MX_H_ */
